added again file exist checking due to worker and agent produces irrelevant logs

+++ stabilizer/src/main/java/com/hazelcast/stabilizer/agent/workerjvm/WorkerJvmLauncher.java
@@ -137,6 +137,10 @@ public class WorkerJvmLauncher {
 
     private void copyResourcesToWorkerId(String workerId) throws IOException {
         final String testSuiteId = agent.getTestSuite().id;
+        if (!new File(WORKERS_PATH + testSuiteId + "/upload/").exists()) {
+            log.info("Skipping copy, since no copy file at the agent");
+            return;
+        }
         String cpCommand = format("cp -rfv %s/%s/upload/* %s/%s/%s/",
                 WORKERS_PATH,
                 testSuiteId,
